Create Day18.js
This problem took me a while, but I finally got it to work. Definitely not the most efficient and very error-prone... I spent a lot of time debugging.
Part 2 was fairly simple although I had to change my calcMagnitude function to return a parseInt version of the result so that I could properly compare number values.